requirements.txt
flask==3.1.0
gunicorn==23.0.0
torch==2.1.0
torchvision==0.15.0
numpy==1.24.3
pillow==10.4.0
